requirements.txt
streamlit
opencv-python-headless
numpy
scikit-image
deepface
mediapipe
matplotlib
Pillow